+++ src/main/java/one/d4d/signsaboteur/itsdangerous/crypto/Signers.java
@@ -1,5 +1,5 @@
 package one.d4d.signsaboteur.itsdangerous.crypto;
 
 public enum Signers {
-    DANGEROUS, EXPRESS, OAUTH, TORNADO, RUBY, JWT, NIMBUSDS, UNKNOWN
+    DANGEROUS, EXPRESS, OAUTH, TORNADO, RUBY, JWT, NIMBUSDS, UNKNOWN, ENCRYPTION
 }